# perf: Speed up cross-file link resolution