--- a/Marksman/Marksman.fsproj
+++ b/Marksman/Marksman.fsproj
@@ -25,6 +25,8 @@
     </Target>
     <ItemGroup>
         <Compile Include="Misc.fs"/>
+        <Compile Include="Paths.fsi"/>
+        <Compile Include="Paths.fs"/>
         <Compile Include="GitIgnore.fs"/>
         <Compile Include="Config.fs"/>
         <Compile Include="Text.fs"/>