… Benchmarks/Benchmarks.fsproj
… Benchmarks/Program.fs
… Marksman/Names.fs
… Marksman/Names.fsi
… Marksman/Paths.fs
… Marksman/Paths.fsi
… Marksman/Refactor.fs
… Marksman/Refs.fs
… Marksman/Server.fs
… Marksman/Symbols.fs
… Marksman/Toc.fs
… Marksman/Workspace.fs
… Marksman/Workspace.fsi
… Tests/ComplTests.fs
… Tests/DiagTest.fs
… Tests/Helpers.fs
… Tests/MiscTests.fs
… Tests/ParserTests.fs
… Tests/PathsTests.fs
… Tests/Program.fs
… Tests/RefsTests.fs
… Tests/SematoTests.fs
… Tests/ServerTests.fs
… Tests/StateTests.fs
… Tests/Tests.fsproj
… Tests/WorkspaceTest.fs