@artempyanykh 2023-05-29
Ok, this looks much better now: 
* goto def is constant time 
* find refs is O(n_docs * n_refs_per_doc) vs O(n^2 * n_refs_per_doc) before.

Which gives a nice 180x (20s vs 115ms) improvement for the 250 docs benchmark.

```
|       Method | FolderSize |           Mean |       Error |      StdDev |
|------------- |----------- |---------------:|------------:|------------:|
|  gotoDefTime |         10 |       1.180 us |   0.0034 us |   0.0032 us |
| findRefsTime |         10 |     147.967 us |   0.7200 us |   0.6735 us |
|  gotoDefTime |         50 |       1.294 us |   0.0013 us |   0.0011 us |
| findRefsTime |         50 |   4,040.372 us |   4.7052 us |   4.4013 us |
|  gotoDefTime |        250 |       1.433 us |   0.0022 us |   0.0019 us |
| findRefsTime |        250 | 115,387.334 us | 118.0649 us | 104.6614 us |
```